+++ app/src/main/java/com/cray/software/justreminder/MainActivity.java
@@ -930,11 +930,11 @@ public class MainActivity extends AppCompatActivity
                 int month;
                 int year;
                 int repCode;
-                int repTime;
+                long repTime;
                 int repCount;
                 repCode = c.getInt(c.getColumnIndex(Constants.COLUMN_REPEAT));
                 repCount = c.getInt(c.getColumnIndex(Constants.COLUMN_REMINDERS_COUNT));
-                repTime = c.getInt(c.getColumnIndex(Constants.COLUMN_REMIND_TIME));
+                repTime = c.getLong(c.getColumnIndex(Constants.COLUMN_REMIND_TIME));
                 day = c.getInt(c.getColumnIndex(Constants.COLUMN_DAY));
                 month = c.getInt(c.getColumnIndex(Constants.COLUMN_MONTH));
                 year = c.getInt(c.getColumnIndex(Constants.COLUMN_YEAR));
@@ -1247,7 +1247,7 @@ public class MainActivity extends AppCompatActivity
                 int month = c.getInt(c.getColumnIndex(Constants.COLUMN_MONTH));
                 int year = c.getInt(c.getColumnIndex(Constants.COLUMN_YEAR));
                 int repCode = c.getInt(c.getColumnIndex(Constants.COLUMN_REPEAT));
-                int repTime = c.getInt(c.getColumnIndex(Constants.COLUMN_REMIND_TIME));
+                long repTime = c.getLong(c.getColumnIndex(Constants.COLUMN_REMIND_TIME));
                 int isDone = c.getInt(c.getColumnIndex(Constants.COLUMN_IS_DONE));
                 double lat = c.getDouble(c.getColumnIndex(Constants.COLUMN_LATITUDE));
                 double lon = c.getDouble(c.getColumnIndex(Constants.COLUMN_LONGITUDE));
